server/src/requirements.txt
aiofiles==23.2.1
aiosqlite==0.19.0
annotated-types==0.5.0
anyio==3.7.1
asyncpg==0.28.0
bcrypt==4.0.1
click==8.1.7
databases==0.8.0
ecdsa==0.18.0
fastapi==0.103.2
h11==0.14.0
idna==3.4
passlib==1.7.4
psycopg2
pyasn1==0.5.0
pydantic==2.4.2
pydantic_core==2.10.1
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2023.3.post1
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.49
starlette==0.27.0
typing_extensions==4.8.0
uvicorn==0.23.2
